test: move isolated test to use examples/starter instead of examples/basic
